requirements.txt
Flask==3.1.0
langchain==0.3.17
langchain_community==0.3.16
numpy==2.2.2
protobuf==5.29.3
sentence_transformers==3.4.1
sqlparse==0.5.3
transformers==4.48.2
faiss-cpu
google-cloud-bigquery